Move getInputClasses solution to FieldIcons
