+++ libs/ui/src/components/field-icons/field-icons.module.css
@@ -1,19 +1,35 @@
 .icons {
   position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  width: 100%;
+  height: 100%;
   display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0 1rem;
+}
+
+.group {
+  display: flex;
+  align-items: center;
   gap: 0.5rem;
-}
-
-.iconsLeft {
-  left: 1rem;
-}
-
-.iconsRight {
-  right: 1rem;
 }
 
 .iconError {
   color: var(--color-danger-300);
+}
+
+.singleIconLeft + input {
+  padding-left: 3rem;
+}
+
+.singleIconRight + input {
+  padding-right: 3rem;
+}
+
+.doubleIconRight + input {
+  padding-right: 5rem;
 }
