+++ libs/ui/src/components/password-field/password-field.module.css
@@ -41,18 +41,6 @@
   }
 }
 
-.inputError {
+.error {
   border-color: var(--color-danger-300);
-}
-
-.leftIconPadding {
-  padding-left: 3rem;
-}
-
-.rightIconPadding {
-  padding-right: 3rem;
-}
-
-.doubleIconPadding {
-  padding-right: 5rem;
 }
